requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-papermill==3.12.1
papermill==2.6.0
dbt-duckdb==1.10.1
ipykernel==6.29.5